automod: refactor engine test fixture in to exported code
