+++ automod/rules/misleading_test.go
@@ -15,7 +15,7 @@ import (
 func TestMisleadingURLPostRule(t *testing.T) {
 	assert := assert.New(t)
 
-	engine := engineFixture()
+	engine := automod.EngineTestFixture()
 	am1 := automod.AccountMeta{
 		Identity: &identity.Identity{
 			DID:    syntax.DID("did:plc:abc111"),
@@ -50,7 +50,7 @@ func TestMisleadingURLPostRule(t *testing.T) {
 func TestMisleadingMentionPostRule(t *testing.T) {
 	assert := assert.New(t)
 
-	engine := engineFixture()
+	engine := automod.EngineTestFixture()
 	am1 := automod.AccountMeta{
 		Identity: &identity.Identity{
 			DID:    syntax.DID("did:plc:abc111"),
